+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Add support for changing the time zone as part of the first-time setup.
 - Add desktop shortcut for first-time setup and only have it run after the first login.
 - Change the user account to require a password when using 'sudo'.
+- Add an image gallery application for Cinnamon (Pix) and KDE Plasma (Gwenview).
 
 ## 2.2.0